# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WALAB API Documentation",
        "description": "Documentation technique complète pour l'application Walab. Utilisez le bouton 'Authorize' pour tester les routes protégées.",
        "contact": {
            "email": "contact@walab.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/examen/get": {
            "get": {
                "tags": [
                    "Examens"
                ],
                "summary": "Détail d'un examen",
                "description": "Récupère les détails d'un examen spécifique par son ID.",
                "operationId": "6b137e49a8c5289a3ebe8c643c0ec183",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID de l'examen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Examen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "ID manquant ou invalide"
                    }
                }
            }
        },
        "/api/examen/create": {
            "post": {
                "tags": [
                    "Examens"
                ],
                "summary": "Créer un examen",
                "description": "Ajoute un nouvel examen à la base de données. (Admin/Labo)",
                "operationId": "574ab095d4daa895d5376aaad4e17565",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "price",
                                    "laboratorie_id"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Groupe Sanguin"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 2500
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Détermination du groupe sanguin"
                                    },
                                    "laboratorie_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "url_image"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Examen créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "create Examen"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Examen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/examen/update": {
            "post": {
                "tags": [
                    "Examens"
                ],
                "summary": "Mettre à jour un examen",
                "operationId": "2c9eb2592d429e554b2f401242e9f25a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "label": {
                                        "type": "string",
                                        "example": "Nouveau Nom"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Examen mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "update Examen"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Examen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/examen/delete": {
            "post": {
                "tags": [
                    "Examens"
                ],
                "summary": "Supprimer un examen",
                "operationId": "ac9f76077ba26540cc8af90a52b18f5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Examen supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "delete Examen"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Examen": {
                "title": "Examen Médical",
                "description": "Détail d'une analyse médicale disponible",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "label": {
                        "type": "string",
                        "example": "NFS (Numération Formule Sanguine)"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 4500
                    },
                    "description": {
                        "type": "string",
                        "example": "Analyse complète du sang..."
                    },
                    "icon": {
                        "type": "string",
                        "example": "examen/nfs.png"
                    },
                    "laboratorie_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "isactive": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'accès ici (sans le préfixe 'Bearer')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}